pyproject.toml
[project]
name = "portfolio-web"
version = "0.1.0"
description = "A personal portfolio website built with Flask, CSS, and GSAP"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.1",
    "python-dotenv>=1.1.1",
]
